Fixes for new MAC environment
git-svn-id: https://forge.ipsl.fr/ioserver/svn/XIOS2/trunk@2470 d613d254-d18a-4304-b296-5125d4c4d670

+++ src/attribute_map.cpp
@@ -380,7 +380,7 @@ namespace xios
            if (it->second->isPublic())
            {
              oss << ", " << it->second->getName() << "_";
-             if (oss.tellp() - startPos > 90)
+             if ((long)oss.tellp() - startPos > 90)
              {
                oss << "  &" << iendl;
                startPos = oss.tellp();
@@ -430,7 +430,7 @@ namespace xios
            if (it->second->isPublic())
            {
              oss << ", " << it->second->getName() << "_";
-             if (oss.tellp() - startPos > 90)
+             if ((long)oss.tellp() - startPos > 90)
              {
                oss << "  &" << iendl;
                startPos = oss.tellp();
@@ -480,7 +480,7 @@ namespace xios
            if (it->second->isPublic())
            {
              oss << ", " << it->second->getName() << "_";
-             if (oss.tellp() - startPos > 90)
+             if ((long)oss.tellp() - startPos > 90)
              {
                oss << "  &" << iendl;
                startPos = oss.tellp();
@@ -530,7 +530,7 @@ namespace xios
            if (it->second->isPublic())
            {
              oss << ", " << it->second->getName();
-             if (oss.tellp() - startPos > 90)
+             if ((long)oss.tellp() - startPos > 90)
              {
                oss << "  &" << iendl;
                startPos = oss.tellp();
@@ -565,7 +565,7 @@ namespace xios
            if (it->second->isPublic())
            {
              oss << ", " << it->second->getName();
-             if (oss.tellp() - startPos > 90)
+             if ((long)oss.tellp() - startPos > 90)
              {
                oss << "  &" << iendl;
                startPos = oss.tellp();
@@ -592,7 +592,7 @@ namespace xios
            if (it->second->isPublic())
            {
              oss << ", " << it->second->getName();
-             if (oss.tellp() - startPos > 90)
+             if ((long)oss.tellp() - startPos > 90)
              {
                oss << "  &" << iendl;
                startPos = oss.tellp();
@@ -627,7 +627,7 @@ namespace xios
            if (it->second->isPublic())
            {
              oss << ", " << it->second->getName();
-             if (oss.tellp() - startPos > 90)
+             if ((long)oss.tellp() - startPos > 90)
              {
                oss << "  &" << iendl;
                startPos = oss.tellp();
@@ -654,7 +654,7 @@ namespace xios
            if (it->second->isPublic())
            {
              oss << ", " << it->second->getName();
-             if (oss.tellp() - startPos > 90)
+             if ((long)oss.tellp() - startPos > 90)
              {
                oss << "  &" << iendl;
                startPos = oss.tellp();
@@ -689,7 +689,7 @@ namespace xios
            if (it->second->isPublic())
            {
              oss << ", " << it->second->getName();
-             if (oss.tellp() - startPos > 90)
+             if ((long)oss.tellp() - startPos > 90)
              {
                oss << "  &" << iendl;
                startPos = oss.tellp();
@@ -716,7 +716,7 @@ namespace xios
            if (it->second->isPublic())
            {
              oss << ", " << it->second->getName();
-             if (oss.tellp() - startPos > 90)
+             if ((long)oss.tellp() - startPos > 90)
              {
                oss << "  &" << iendl;
                startPos = oss.tellp();
@@ -754,7 +754,7 @@ namespace xios
            if (it->second->isPublic())
            {
              oss << ", " << it->second->getName();
-             if (oss.tellp() - startPos > 90)
+             if ((long)oss.tellp() - startPos > 90)
              {
                oss << "  &" << iendl;
                startPos = oss.tellp();
@@ -781,7 +781,7 @@ namespace xios
            if (it->second->isPublic())
            {
              oss << ", " << it->second->getName();
-             if (oss.tellp() - startPos > 90)
+             if ((long)oss.tellp() - startPos > 90)
              {
                oss << "  &" << iendl;
                startPos = oss.tellp();
@@ -819,7 +819,7 @@ namespace xios
            if (it->second->isPublic())
            {
              oss << ", " << it->second->getName();
-             if (oss.tellp() - startPos > 90)
+             if ((long)oss.tellp() - startPos > 90)
              {
                oss << "  &" << iendl;
                startPos = oss.tellp();
@@ -846,7 +846,7 @@ namespace xios
            if (it->second->isPublic())
            {
              oss << ", " << it->second->getName();
-             if (oss.tellp() - startPos > 90)
+             if ((long)oss.tellp() - startPos > 90)
              {
                oss << "  &" << iendl;
                startPos = oss.tellp();
@@ -884,7 +884,7 @@ namespace xios
            if (it->second->isPublic())
            {
              oss << ", " << it->second->getName();
-             if (oss.tellp() - startPos > 90)
+             if ((long)oss.tellp() - startPos > 90)
              {
                oss << "  &" << iendl;
                startPos = oss.tellp();
